+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug: fix file storage new mutation as deleted with put fail error 
 - Enh #23: extend storage interface with modify
 - Enh #56: content hash added to gmail message objects
+- Enh #61: separated OAuth token storage - RENEWAL OF OAUTH AUTHENTICATION IS REQUIRED 
 
 # 0.8.0